Move utility functions in the api file
